Reorganize DAO-packages
