+++ app/build.gradle
@@ -40,6 +40,7 @@ android {
 
 greendao {
     schemaVersion 5
+    daoPackage 'fr.gaulupeau.apps.Poche.data.dao'
 }
 
 dependencies {
